Domain and Transport templates platforms support

+++ dds/DCPS/transport/framework/TransportRegistry.cpp
@@ -746,9 +746,9 @@ TransportRegistry::create_transport_template_instance(DDS::DomainId_t domain, co
             size_t pos = addr.find_last_of(":");
             if (pos == OPENDDS_STRING::npos) {
               // See 9.6.1.3 in the RTPS 2.2 protocol specification.
-              const uint16_t PB = 7400;
-              const uint16_t DG = 250;
-              const uint16_t D2 = 1;
+              const int PB = 7400;
+              const int DG = 250;
+              const int D2 = 1;
               int rtpsPort = PB + DG * domain + D2;
               addr += ":" + to_dds_string(rtpsPort);
             } else {
